+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #27 Integrate dispatch workflow from senaite.core
+- #26 Disallow to modify portal content when stored/booked out samples
 - #25 Remove booked out samples from container
 - #23 Allow storage container to be booked out
 - #22 Allow to move containers